% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markov.R
\name{Markov1}
\alias{Markov1}
\title{Discrete time Markov model class}
\description{
Discrete time Markov model class

Discrete time Markov model class
}
\details{
Class for the discrete time Markov model, calculating reserves
(Deckungskapital (DK) in German \code{reserve}) and cashflows (CF,
\code{cashflow})given a transition matrix (P, \code{transition_matrix}),
prenumerando (pre, \code{prenumarando}) and postnumerando (post,
\code{postnumerando}) benefits for each time point. You can also specify a
discount rate (v, \code{discount_rate})
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n_states}}{\href{1}{integer} Number of distinct states in the Markov
model}

\item{\code{transition_matrix}}{\link{array}(n_states, n_states, max_time) Transition
matrix of the Markov model for each time point \code{t} in the first
dimension and states \code{i} and \code{j} in the second and third
dimensions}

\item{\code{prenumerando}}{\link{matrix}(max_time, n_states) Prenumerando benefits
for each time \code{t} and state \code{i}}

\item{\code{postnumerando}}{\link{array}(n_states, n_states, max_time) Postnumerando
benefits for each time \code{t} and states \code{i} and \code{j}}

\item{\code{discount_factor}}{\href{time}{numeric} Discount factors to apply to
cashflows to calculate present values}

\item{\code{reserve}}{\link{matrix}(max_time, n_states) Mathematical reserve (or
Deckungskapital (DK) in German) for each time \code{t} and state
\code{i}}

\item{\code{cashflow}}{\link{matrix}(max_time, n_states) Cashflows for each time
\code{t} and state \code{i}}

\item{\code{start}}{\href{1}{integer} Start time \code{t} for calculations}

\item{\code{stop}}{\href{1}{integer} End time \code{t} for calculations}

\item{\code{max_time}}{\href{1}{integer} Maximum amount of time steps \code{t} in the
model}

\item{\code{state_names}}{\href{n_states}{character} Names for the states in the
model}

\item{\code{is_calculated_reserve}}{\href{1}{logical} Indicator whether model
technical results have been calculated with most recent configuration}

\item{\code{is_calculated_cashflow}}{\href{1}{logical} Indicator whether model
cashflows have been calculated with the most recent configuration}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Markov1-new}{\code{Markov1$new()}}
\item \href{#method-Markov1-set_discount_factor}{\code{Markov1$set_discount_factor()}}
\item \href{#method-Markov1-set_transition_matrix}{\code{Markov1$set_transition_matrix()}}
\item \href{#method-Markov1-set_prenumerando}{\code{Markov1$set_prenumerando()}}
\item \href{#method-Markov1-set_postnumerando}{\code{Markov1$set_postnumerando()}}
\item \href{#method-Markov1-calculate_reserve}{\code{Markov1$calculate_reserve()}}
\item \href{#method-Markov1-calculate_cashflow}{\code{Markov1$calculate_cashflow()}}
\item \href{#method-Markov1-get_reserve}{\code{Markov1$get_reserve()}}
\item \href{#method-Markov1-get_cashflow}{\code{Markov1$get_cashflow()}}
\item \href{#method-Markov1-plot_reserve}{\code{Markov1$plot_reserve()}}
\item \href{#method-Markov1-plot_cashflow}{\code{Markov1$plot_cashflow()}}
\item \href{#method-Markov1-reset_calculation}{\code{Markov1$reset_calculation()}}
\item \href{#method-Markov1-transform_for_plot}{\code{Markov1$transform_for_plot()}}
\item \href{#method-Markov1-print_output}{\code{Markov1$print_output()}}
\item \href{#method-Markov1-clone}{\code{Markov1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-new"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-new}{}}}
\subsection{Method \code{new()}}{
Create a new object of the \code{Markov} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$new(n_states)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_states}}{\href{1}{integer} The number of states in the Markov model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{Markov1} An object of the \code{Markov} class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-set_discount_factor"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-set_discount_factor}{}}}
\subsection{Method \code{set_discount_factor()}}{
Set the discount rate \code{discount_rate} for the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$set_discount_factor(rate, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rate}}{\href{time}{numeric} Set discount rates to use for calculating
discount rates under discrete annual compounding convention. Scalar values
will be recycled for all times \code{t}}

\item{\code{verbose}}{\href{1}{logical} If \code{TRUE} print the resulting discount
factor, if \code{FALSE} set discount factor silently}

\item{\code{verbose}}{\href{1}{logical} If \code{TRUE} show verbose output for each
iteration in the calculation. If \code{FALSE} do the calculation silently}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sets the \code{discount_factor} field and returns the model
invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-set_transition_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-set_transition_matrix}{}}}
\subsection{Method \code{set_transition_matrix()}}{
Set the transition matrix \code{transition_matrix} for the
model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$set_transition_matrix(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{\link{array}(n_states, n_states, max_time) Transition matrix of the
Markov model for each time point \code{t} in the first dimension and states
\code{i} and \code{j} in the second and third dimensions. If
\link{matrix}(n_states, n_states) it will be recycled for all time steps}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sets the \code{transition_matrix} field and returns the model
invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-set_prenumerando"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-set_prenumerando}{}}}
\subsection{Method \code{set_prenumerando()}}{
Set the prenumerando benefits \code{prenumerando} for the
model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$set_prenumerando(pre)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pre}}{\link{matrix}(max_time, n_states) Prenumerando benefits for each time
\code{t} and state \code{i}. If \href{n_states}{numeric} it will be recycled for
all time steps}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sets the \code{prenumerando} field and returns the model
invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-set_postnumerando"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-set_postnumerando}{}}}
\subsection{Method \code{set_postnumerando()}}{
Set the postnumerando benefits \code{postnumerando} for the
model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$set_postnumerando(post)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{post}}{\link{array}(n_states, n_states, max_time) Postnumerando benefits for
each time \code{t} and states \code{i}, \code{j}. If \link{matrix}(n_states,
n_states) it will be recycled for all time steps}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sets the \code{postnumerando} field and returns the model
invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-calculate_reserve"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-calculate_reserve}{}}}
\subsection{Method \code{calculate_reserve()}}{
Calculate the mathematical reserve (Deckungskapital (DK) in
German)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$calculate_reserve(start, stop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\href{1}{integer} Start time for calculation}

\item{\code{stop}}{\href{1}{integer} Stop time for calculation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Calculates the mathematical reserves, stores them in the
\code{reserve} field and returns the model invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-calculate_cashflow"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-calculate_cashflow}{}}}
\subsection{Method \code{calculate_cashflow()}}{
Calculate the model cashflows
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$calculate_cashflow(start, stop, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\href{1}{integer} Start time for calculation}

\item{\code{stop}}{\href{1}{integer} Stop time for calculation}

\item{\code{verbose}}{\href{1}{logical} If \code{TRUE} print the resulting discount
factor, if \code{FALSE} set discount factor silently}

\item{\code{verbose}}{\href{1}{logical} If \code{TRUE} show verbose output for each
iteration in the calculation. If \code{FALSE} do the calculation silently}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Calculates the model cashflows, stores them in the
\code{cashflow} field and returns the model invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-get_reserve"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-get_reserve}{}}}
\subsection{Method \code{get_reserve()}}{
Get mathematical reserves between two times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$get_reserve(start, stop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\href{1}{integer} Start time for calculation}

\item{\code{stop}}{\href{1}{integer} Stop time for calculation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{matrix}(start - stop, n_states) Mathematical reserve from
times \code{start} to \code{stop}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-get_cashflow"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-get_cashflow}{}}}
\subsection{Method \code{get_cashflow()}}{
Get model cashflows between two times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$get_cashflow(start, stop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\href{1}{integer} Start time for calculation}

\item{\code{stop}}{\href{1}{integer} Stop time for calculation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{matrix}(start - stop, n_states) Model cashflows from times
\code{start} to \code{stop}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-plot_reserve"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-plot_reserve}{}}}
\subsection{Method \code{plot_reserve()}}{
Plot a line graph of reserves
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$plot_reserve(start, stop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\href{1}{integer} Start time for calculation}

\item{\code{stop}}{\href{1}{integer} Stop time for calculation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[ggplot2:ggplot]{ggplot2::ggplot} \link{ggplot2} plot of reserves
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-plot_cashflow"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-plot_cashflow}{}}}
\subsection{Method \code{plot_cashflow()}}{
Plot a column graph of cashflows
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$plot_cashflow(start, stop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{\href{1}{integer} Start time for calculation}

\item{\code{stop}}{\href{1}{integer} Stop time for calculation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[ggplot2:ggplot]{ggplot2::ggplot} \link{ggplot2} plot of cashflows
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-reset_calculation"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-reset_calculation}{}}}
\subsection{Method \code{reset_calculation()}}{
Reset the state of the model as non-calculated
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$reset_calculation()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Sets \code{is_calculated_reserve} and
\code{is_calculated_cashflow} fields to \code{FALSE} and returns the
model invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-transform_for_plot"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-transform_for_plot}{}}}
\subsection{Method \code{transform_for_plot()}}{
Transform data for plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$transform_for_plot(x, values_to, names_to = "state")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\link{data.frame} Data set to transform}

\item{\code{values_to}}{\href{1}{character} Name of the value column, passed to
\code{\link[tidyr]{tidyr}}}

\item{\code{names_to}}{\href{1}{character} Name of the state column, passed to
\code{\link[tidyr]{tidyr}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
[\link[tibble]{tibble}] Data set in appropriate format for
\link{ggplot2} plots
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-print_output"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-print_output}{}}}
\subsection{Method \code{print_output()}}{
Format output in a user-friendly \link{data.frame}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$print_output(output, start = self$max_time, stop = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{\link{matrix}(max_time, n_states) Output to be formatted}

\item{\code{start}}{\href{1}{integer} Start time for calculation}

\item{\code{stop}}{\href{1}{integer} Stop time for calculation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{data.frame} Model output with \code{time} and values for each
state
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Markov1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Markov1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Markov1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
