% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovlv-package.R
\docType{package}
\name{markovlv-package}
\alias{markovlv}
\alias{markovlv-package}
\title{markovlv: Discrete time Markov class}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Benedikt Schamberger \email{schamberger.benedikt@gmail.com}

}
\keyword{internal}
